+++ packages/frameworks/vue/src/select/select-item-indicator.tsx
@@ -3,7 +3,7 @@ import { ark, type HTMLArkProps } from '../factory'
 import { useSelectContext } from './select-context'
 import { useSelectItemContext } from './select-item-context'
 
-export type SelectItemIndicatorProps = HTMLArkProps<'div'>
+export interface SelectItemIndicatorProps extends HTMLArkProps<'div'> {}
 
 export const SelectItemIndicator = defineComponent({
   name: 'SelectItemIndicator',