+++ packages/frameworks/vue/src/select/select-trigger.tsx
@@ -2,7 +2,7 @@ import { defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
 import { useSelectContext } from './select-context'
 
-export type SelectTriggerProps = HTMLArkProps<'button'>
+export interface SelectTriggerProps extends HTMLArkProps<'button'> {}
 
 export const SelectTrigger = defineComponent({
   name: 'SelectTrigger',