+++ packages/frameworks/vue/src/select/use-select.ts
@@ -3,18 +3,25 @@ import * as select from '@zag-js/select'
 import { normalizeProps, useMachine, type PropTypes } from '@zag-js/vue'
 import { computed, type ComputedRef } from 'vue'
 import { useEnvironmentContext } from '../environment'
-import type { Optional } from '../types'
+import type { CollectionItem, Optional } from '../types'
 import { generateEventMap, useId } from '../utils'
 import { emits } from './select.props'
 
-export interface UseSelectProps
-  extends CollectionOptions,
-    Omit<Optional<select.Context, 'id'>, 'collection'> {
-  modelValue?: select.Context['value']
+export interface UseSelectProps<T extends CollectionItem>
+  extends CollectionOptions<T>,
+    Omit<Optional<select.Context<T>, 'id'>, 'collection'> {
+  modelValue?: select.Context<T>['value']
 }
-export type UseSelectReturn = ComputedRef<select.Api<PropTypes>>
 
-export const useSelect = (props: UseSelectProps, emit: CallableFunction): UseSelectReturn => {
+export interface UseSelectReturn<T extends CollectionItem>
+  extends ComputedRef<select.Api<PropTypes, T>> {}
+
+export const useSelect = <T extends CollectionItem>(
+  props: UseSelectProps<T>,
+  emit: CallableFunction,
+): UseSelectReturn<T> => {
+  const getRootNode = useEnvironmentContext()
+
   const { items, itemToString, itemToValue, isItemDisabled, ...selectProps } = props
   const context = computed(() => {
     const { modelValue, ...rest } = selectProps
@@ -24,7 +31,6 @@ export const useSelect = (props: UseSelectProps, emit: CallableFunction): UseSel
     }
   })
   const collection = select.collection({ items, itemToString, itemToValue, isItemDisabled })
-  const getRootNode = useEnvironmentContext()
   const eventMap = generateEventMap(emits, emit)
 
   const [state, send] = useMachine(